Scaling
Have included lambda functions to scale the height map.

I think tomorrow I'll aim to create a jupyter notebook to run the full code in, so that I dont have to load in the data everytime I want to make a small tweak. Then I'll start on creating coordinate systems (geocoords to image coords) that I can then use with the hexagons.