Testing code
I've changed the notebook to allow me to test the coordinate transforms and the hexagons.

As for the hexagons, they work very well, and I'm able to generate adjacent hexagons at given positions, and cull the image based on being internal or external to the hexagons.

The CoordinateTransform isn't properly calibrated. I want to see if I can do it from the file metadata, but that will require some more research.