Fixed weighting function
I've adjusted the default weighting function in the grid interpolation, so that it can't go negative on the boundary x,y in [-1,1]. Before, this was possible at (1,1), etc, which was resulting in spiky terrain near elevation changes - most noticable at coasts.

Adding some constant to the given cos(x)cos(y) function will result in a smoothing function.