Coaster STL files
I've included a new function in HexGrid to allow for the selection of corner indices. This has allowed me to raise them, which would allow me to put a raised platform above the topography.

I've also changed main.py as the loading of files needs seperate folder seperators in mac and windows :(